fix(tests): restore suites after the processor architecture shift (#50)
Bring the restored test suite back into alignment with the processor-era
database APIs by expanding the query builder, fixing provenance array
handling, and sanding down harness mismatches across the workspace.

sinex-db (query builder and queries):
- add `QueryParam::Float`, `OptionalFloat`, `WhereCondition::NullCheck`,
  `where_is_null`, and `where_is_not_null` in
  `crate/sinex-db/src/query_builder.rs`, so db helpers can represent
  current SQL shapes without smuggling NULL logic through string values
- add `OptionalUlidArray` support and use it in event insert paths, which
  fixes nullable `source_event_ids` handling for raw vs synthesized event
  provenance
- correct `source_event_ids` casts to `ulid[]` in
  `crate/sinex-db/src/queries/events.rs` and
  `crate/sinex-db/src/events_migrated.rs`, eliminating the UUID/ULID
  mismatch that was still breaking restored suites and property tests
- add new blob and analytics query helpers, including blob lookups and
  storage stats in `crate/sinex-db/src/queries/artifacts.rs` plus event
  count, top-command, and heartbeat queries in
  `crate/sinex-db/src/queries/events.rs`

runtime follow-through:
- adapt `crate/sinex-annex/src/blob_manager.rs`, preflight code, service
  crates, and satellite code to use the new helper surface and current
  source constants
- wire PostgreSQL extensions through NixOS configuration and retire the
  standalone ULID-extension migration, so test environments match the
  processor-era database assumptions

test harness and suites:
- add mock scanner/event-source helpers, explicit wait loops, pool-borrow
  fixes, Redis reply parsing fixes, helper-struct realignment, config/ID
  refreshes, and validator lifetime fixes across `test/common/*`
- port integration, property, system, adversarial, and performance suites
  to the current `CoreError`, `TestContext`, query-builder, and Redis
  APIs, including checkpoint, corruption, PEL recovery, provenance, and
  reliability coverage
- make `sinex_test` return `Box<dyn Error + Send + Sync>` and normalize
  proptest error conversion in `crate/sinex-test-macros/src/lib.rs`, so
  property-heavy and long-running suites are safe to spawn under Tokio

+++ crate/sinex-db/src/queries/operations.rs
@@ -368,7 +368,7 @@ impl OperationQueries {
         }
 
         if let Some(lim) = limit {
-            builder = builder.limit(lim as usize);
+            builder = builder.limit(lim);
         } else {
             builder = builder.limit(1000);
         }
